requirements.txt
customtkinter==5.1.3
Pillow==9.5.0

